requirements.txt

django-crispy-forms==1.6.0
python-dateutil
